1.02

+++ src/components/subItem/SubItem.js
@@ -10,23 +10,36 @@ export const SubItem = (props) => {
     }
 
     const onDelete = (e) => {
-        for (let i = 0; i < data.days.length; i++) {
-            for (let u = 0; u < data.days[i].subItems.length; u++ ) {
-                if (data.days[i].subItems[u].id == id) {
-                    const selectedDay = data.days[i].day;
-                    //Code
-                    const updatedDays = data.days.map((item) => {
-                        if (item.day === selectedDay) {
-                            return { ...item, subItems: data.days[i].subItems.filter(itm => itm.id != id) };
+        const Swal = require("sweetalert2");
+        Swal.fire({
+            title: "Вы уверенны что хотите удалить?",
+            text: "Вы не сможете вернуть обратно",
+            icon: "warning",
+            showCancelButton: true,
+            confirmButtonColor: "#3085d6",
+            cancelButtonColor: "#d33",
+            confirmButtonText: "Да, удалить!"
+          }).then((result) => {
+            if (result.isConfirmed) {
+                for (let i = 0; i < data.days.length; i++) {
+                    for (let u = 0; u < data.days[i].subItems.length; u++ ) {
+                        if (data.days[i].subItems[u].id == id) {
+                            const selectedDay = data.days[i].day;
+                            //Code
+                            const updatedDays = data.days.map((item) => {
+                                if (item.day === selectedDay) {
+                                    return { ...item, subItems: data.days[i].subItems.filter(itm => itm.id != id) };
+                                }
+                                return item;
+                            });
+                            localStorage.removeItem("data");
+                            localStorage.setItem("data", JSON.stringify({...data, days: updatedDays}));
+                            setData(JSON.parse(localStorage.getItem("data")));
                         }
-                        return item;
-                    });
-                    localStorage.removeItem("data");
-                    localStorage.setItem("data", JSON.stringify({...data, days: updatedDays}));
-                    setData(JSON.parse(localStorage.getItem("data")));
+                    }
                 }
             }
-        }
+          });
     }
 
     return (
